+++ CHANGELOG.md
@@ -39,6 +39,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Features
 
+* (x/collection) [\#1038](https://github.com/Finschia/finschia-sdk/pull/1038) Update x/collection cli for hackwasm2023
+
 ### Improvements
 
 ### Bug Fixes